requirements.txt
selenium>=4.10.0
google-cloud-vision>=2.7.3
python-dotenv>=0.21.0
webdriver-manager>=3.8.0
pillow>=9.0.0
pytest>=7.0.0
requests>=2.28.0
beautifulsoup4>=4.11.0
lxml>=4.9.0